# simplyblock_web/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v2/clusters/": {
      "get": {
        "summary": "Clusters:List",
        "operationId": "clusters_list_api_v2_clusters__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClusterDTO"
                  },
                  "title": "Response Clusters List Api V2 Clusters  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Clusters:Create",
        "operationId": "clusters_create_api_v2_clusters__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterParams"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/": {
      "get": {
        "summary": "Clusters:Detail",
        "operationId": "clusters_detail_api_v2_clusters__cluster_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Clusters:Update",
        "operationId": "clusters_update_api_v2_clusters__cluster_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatableClusterParameters"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clusters:Delete",
        "operationId": "clusters_delete_api_v2_clusters__cluster_id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/capacity": {
      "get": {
        "summary": "Clusters:Capacity",
        "operationId": "clusters_capacity_api_v2_clusters__cluster_id__capacity_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          },
          {
            "name": "history",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/iostats": {
      "get": {
        "summary": "Clusters:Iostats",
        "operationId": "clusters_iostats_api_v2_clusters__cluster_id__iostats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          },
          {
            "name": "history",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "History"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/logs": {
      "get": {
        "summary": "Clusters:Logs",
        "operationId": "clusters_logs_api_v2_clusters__cluster_id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/start": {
      "post": {
        "summary": "Clusters:Start",
        "operationId": "clusters_start_api_v2_clusters__cluster_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/shutdown": {
      "post": {
        "summary": "Clusters:Shutdown",
        "operationId": "clusters_shutdown_api_v2_clusters__cluster_id__shutdown_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/activate": {
      "post": {
        "summary": "Clusters:Activate",
        "operationId": "clusters_activate_api_v2_clusters__cluster_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/addreplication": {
      "post": {
        "summary": "Clusters:Addreplication",
        "operationId": "clusters_addreplication_api_v2_clusters__cluster_id__addreplication_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ReplicationParams"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/expand": {
      "post": {
        "summary": "Clusters:Expand",
        "operationId": "clusters_expand_api_v2_clusters__cluster_id__expand_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/clusters/{cluster_id}/update": {
      "post": {
        "summary": "Clusters:Upgrade",
        "operationId": "clusters_upgrade_api_v2_clusters__cluster_id__update_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cluster Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_UpdateParams"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapacityStatDTO": {
        "properties": {
          "date": {
            "type": "integer",
            "title": "Date"
          },
          "size_total": {
            "type": "integer",
            "title": "Size Total"
          },
          "size_prov": {
            "type": "integer",
            "title": "Size Prov"
          },
          "size_used": {
            "type": "integer",
            "title": "Size Used"
          },
          "size_free": {
            "type": "integer",
            "title": "Size Free"
          },
          "size_util": {
            "type": "integer",
            "title": "Size Util"
          }
        },
        "type": "object",
        "required": [
          "date",
          "size_total",
          "size_prov",
          "size_used",
          "size_free",
          "size_util"
        ],
        "title": "CapacityStatDTO"
      },
      "ClusterDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "nqn": {
            "type": "string",
            "title": "Nqn"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "read_only",
              "inactive",
              "suspended",
              "degraded",
              "unready",
              "in_activation",
              "in_expansion"
            ],
            "title": "Status"
          },
          "is_re_balancing": {
            "type": "boolean",
            "title": "Is Re Balancing"
          },
          "block_size": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Block Size"
          },
          "distr_ndcs": {
            "type": "integer",
            "title": "Distr Ndcs"
          },
          "distr_npcs": {
            "type": "integer",
            "title": "Distr Npcs"
          },
          "ha": {
            "type": "boolean",
            "title": "Ha"
          },
          "utliziation_critical": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Utliziation Critical"
          },
          "utilization_warning": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Utilization Warning"
          },
          "provisioned_cacacity_critical": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Provisioned Cacacity Critical"
          },
          "provisioned_cacacity_warning": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Provisioned Cacacity Warning"
          },
          "node_affinity": {
            "type": "boolean",
            "title": "Node Affinity"
          },
          "anti_affinity": {
            "type": "boolean",
            "title": "Anti Affinity"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "tls_enabled": {
            "type": "boolean",
            "title": "Tls Enabled"
          },
          "max_fault_tolerance": {
            "type": "integer",
            "title": "Max Fault Tolerance"
          },
          "backup_enabled": {
            "type": "boolean",
            "title": "Backup Enabled"
          },
          "capacity": {
            "$ref": "#/components/schemas/CapacityStatDTO"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "nqn",
          "status",
          "is_re_balancing",
          "block_size",
          "distr_ndcs",
          "distr_npcs",
          "ha",
          "utliziation_critical",
          "utilization_warning",
          "provisioned_cacacity_critical",
          "provisioned_cacacity_warning",
          "node_affinity",
          "anti_affinity",
          "secret",
          "tls_enabled",
          "max_fault_tolerance",
          "backup_enabled",
          "capacity"
        ],
        "title": "ClusterDTO"
      },
      "ClusterParams": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "blk_size": {
            "type": "integer",
            "enum": [
              512,
              4096
            ],
            "title": "Blk Size",
            "default": 512
          },
          "page_size_in_blocks": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Page Size In Blocks",
            "default": 2097152
          },
          "cap_warn": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cap Warn",
            "default": 0
          },
          "cap_crit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cap Crit",
            "default": 0
          },
          "prov_cap_warn": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Prov Cap Warn",
            "default": 0
          },
          "prov_cap_crit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Prov Cap Crit",
            "default": 0
          },
          "distr_ndcs": {
            "type": "integer",
            "title": "Distr Ndcs",
            "default": 1
          },
          "distr_npcs": {
            "type": "integer",
            "title": "Distr Npcs",
            "default": 1
          },
          "distr_bs": {
            "type": "integer",
            "title": "Distr Bs",
            "default": 4096
          },
          "distr_chunk_bs": {
            "type": "integer",
            "title": "Distr Chunk Bs",
            "default": 4096
          },
          "ha_type": {
            "type": "string",
            "enum": [
              "single",
              "ha"
            ],
            "title": "Ha Type",
            "default": "ha"
          },
          "qpair_count": {
            "type": "integer",
            "title": "Qpair Count",
            "default": 256
          },
          "max_queue_size": {
            "type": "integer",
            "title": "Max Queue Size",
            "default": 128
          },
          "inflight_io_threshold": {
            "type": "integer",
            "title": "Inflight Io Threshold",
            "default": 4
          },
          "enable_node_affinity": {
            "type": "boolean",
            "title": "Enable Node Affinity",
            "default": false
          },
          "strict_node_anti_affinity": {
            "type": "boolean",
            "title": "Strict Node Anti Affinity",
            "default": false
          },
          "is_single_node": {
            "type": "boolean",
            "title": "Is Single Node",
            "default": false
          },
          "fabric": {
            "type": "string",
            "title": "Fabric",
            "default": "tcp"
          },
          "cr_name": {
            "type": "string",
            "title": "Cr Name",
            "default": ""
          },
          "cr_namespace": {
            "type": "string",
            "title": "Cr Namespace",
            "default": ""
          },
          "cr_plural": {
            "type": "string",
            "title": "Cr Plural",
            "default": ""
          },
          "cluster_ip": {
            "type": "string",
            "title": "Cluster Ip",
            "default": ""
          },
          "grafana_secret": {
            "type": "string",
            "title": "Grafana Secret",
            "default": ""
          },
          "client_data_nic": {
            "type": "string",
            "title": "Client Data Nic",
            "default": ""
          },
          "max_fault_tolerance": {
            "type": "integer",
            "title": "Max Fault Tolerance",
            "default": 1
          },
          "nvmf_base_port": {
            "type": "integer",
            "title": "Nvmf Base Port",
            "default": 4420
          },
          "rpc_base_port": {
            "type": "integer",
            "title": "Rpc Base Port",
            "default": 8080
          },
          "snode_api_port": {
            "type": "integer",
            "title": "Snode Api Port",
            "default": 50001
          },
          "backup_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackupConfigParams"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ClusterParams"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdatableClusterParameters": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "UpdatableClusterParameters"
      },
...
      "_ReplicationParams": {
        "properties": {
          "snapshot_replication_target_cluster": {
            "type": "string",
            "title": "Snapshot Replication Target Cluster"
          },
          "snapshot_replication_timeout": {
            "type": "integer",
            "title": "Snapshot Replication Timeout",
            "default": 0
          },
          "target_pool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Pool"
          }
        },
        "type": "object",
        "required": [
          "snapshot_replication_target_cluster"
        ],
        "title": "_ReplicationParams"
      },
...
      "_UpdateParams": {
        "properties": {
          "management_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Management Image"
          },
          "spdk_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spdk Image"
          },
          "restart": {
            "type": "boolean",
            "title": "Restart",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "management_image",
          "spdk_image"
        ],
        "title": "_UpdateParams"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}